fix bugs with category
